+++ NEWS.md
@@ -11,10 +11,12 @@ Minor update for fine-tuning internal functions. No new features.
 
 ## Changes
 
-* monotonicity check in Taylor polynomilas now accept minimal inconsistencies
-  (violations of less than .1% of the raw score range)
+* monotonicity check in Taylor polynomials now accept minimal inconsistencies
+  (violations of less than 1% of the raw score range; parameter added to cnorm 
+  and bestModel).
 * The averaging feature has been turned of by default. We have to conduct 
   more research first.
+* deprectated subsampling parameter and according function removed